# openapi.json
{"openapi":"3.1.0","info":{"title":"Spending tracker API","version":"0.1.0"},"servers":[{"url":"/api"}],"paths":{"/":{"get":{"summary":"Get Template","operationId":"get_template__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/auth":{"post":{"summary":"Auth","operationId":"auth_auth_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialsDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseContentDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/greetings":{"get":{"summary":"Greetings","operationId":"greetings_greetings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseContentDTO"}}}}}}},"/registration":{"post":{"summary":"Create User","operationId":"create_user_registration_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialsDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseContentDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/{id}":{"get":{"tags":["User"],"summary":"Get User","operationId":"get_user_user__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}},{"name":"with_relations","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":false,"title":"With Relations"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["User"],"summary":"Update User","operationId":"update_user_user__id__put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["User"],"summary":"Delete User","operationId":"delete_user_user__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}},{"name":"soft_delete","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":true,"title":"Soft Delete"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseContentDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/purchase":{"post":{"tags":["Purchase"],"summary":"Create Purchase","operationId":"create_purchase_purchase_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseCreateDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseContentDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/purchase/{id}":{"get":{"tags":["Purchase"],"summary":"Get Purchase","operationId":"get_purchase_purchase__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Purchase"],"summary":"Delete Purchase","operationId":"delete_purchase_purchase__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}},{"name":"soft_delete","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":true,"title":"Soft Delete"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseContentDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/bank_account":{"post":{"tags":["Bank account"],"summary":"Create Bank Account","operationId":"create_bank_account_bank_account_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BankAccountCreateDTO"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseContentDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Bank account"],"summary":"Get Bank Accounts","operationId":"get_bank_accounts_bank_account_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BankAccountDTO"},"title":"Response Get Bank Accounts Bank Account Get"}}}}}},"delete":{"tags":["Bank account"],"summary":"Delete Bank Accounts","operationId":"delete_bank_accounts_bank_account_delete","parameters":[{"name":"soft_delete","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":true,"title":"Soft Delete"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseContentDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/bank_account/{id}":{"get":{"tags":["Bank account"],"summary":"Get Bank Account","operationId":"get_bank_account_bank_account__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BankAccountDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Bank account"],"summary":"Delete Bank Account","operationId":"delete_bank_account_bank_account__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}},{"name":"soft_delete","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":true,"title":"Soft Delete"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseContentDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BankAccountCreateDTO":{"properties":{"account":{"type":"string","maxLength":20,"minLength":4,"title":"Account","description":"account can contain only letters, numbers and spaces"},"issuer_id":{"type":"integer","maximum":11.0,"minimum":1.0,"title":"Issuer Id","description":"issuer_id must be an existing id in bank_account_issuers table"},"currency_id":{"type":"integer","maximum":4.0,"minimum":1.0,"title":"Currency Id","description":"currency_id must be an existing id in currencies table"},"user_id":{"type":"integer","title":"User Id","description":"user_id must be an existing id in users table"}},"type":"object","required":["account","issuer_id","currency_id","user_id"],"title":"BankAccountCreateDTO"},"BankAccountDTO":{"properties":{"account":{"type":"string","maxLength":20,"minLength":4,"title":"Account","description":"account can contain only letters, numbers and spaces"},"issuer_id":{"type":"integer","maximum":11.0,"minimum":1.0,"title":"Issuer Id","description":"issuer_id must be an existing id in bank_account_issuers table"},"currency_id":{"type":"integer","maximum":4.0,"minimum":1.0,"title":"Currency Id","description":"currency_id must be an existing id in currencies table"},"user_id":{"type":"integer","title":"User Id","description":"user_id must be an existing id in users table"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"},"deleted_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Deleted At"},"purchases":{"items":{"$ref":"#/components/schemas/PurchaseDTO"},"type":"array","title":"Purchases"}},"type":"object","required":["account","issuer_id","currency_id","user_id","id","created_at","updated_at","deleted_at"],"title":"BankAccountDTO"},"CredentialsDTO":{"properties":{"login":{"type":"string","maxLength":20,"minLength":4,"title":"Login","description":"login must be without spaces and can contain only letters and numbers"},"password":{"type":"string","maxLength":20,"minLength":6,"title":"Password","description":"password must be without spaces and can contain only letters and numbers"}},"type":"object","required":["login","password"],"title":"CredentialsDTO"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PurchaseCreateDTO":{"properties":{"cost":{"type":"number","maximum":10000000.0,"minimum":0.01,"title":"Cost","description":"cost must be floating point number rounded to two decimal places"},"account_id":{"type":"integer","title":"Account Id","description":"account_id must be an existing id in bank_accounts table"},"sub_type_id":{"type":"integer","maximum":108.0,"minimum":1.0,"title":"Sub Type Id","description":"sub_type_id must be an existing id in product_sub_types table"}},"type":"object","required":["cost","account_id","sub_type_id"],"title":"PurchaseCreateDTO"},"PurchaseDTO":{"properties":{"cost":{"type":"number","maximum":10000000.0,"minimum":0.01,"title":"Cost","description":"cost must be floating point number rounded to two decimal places"},"account_id":{"type":"integer","title":"Account Id","description":"account_id must be an existing id in bank_accounts table"},"sub_type_id":{"type":"integer","maximum":108.0,"minimum":1.0,"title":"Sub Type Id","description":"sub_type_id must be an existing id in product_sub_types table"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"},"deleted_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["cost","account_id","sub_type_id","id","created_at","updated_at","deleted_at"],"title":"PurchaseDTO"},"ResponseContentDTO":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"},"data":{"anyOf":[{"type":"string"},{"additionalProperties":true,"type":"object"},{"items":{},"type":"array"}],"title":"Data"}},"type":"object","required":["success","message","data"],"title":"ResponseContentDTO"},"UserDTO":{"properties":{"id":{"type":"integer","title":"Id"},"login":{"type":"string","title":"Login"},"hashed_password":{"type":"string","title":"Hashed Password"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"},"deleted_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Deleted At"},"bank_accounts":{"items":{"$ref":"#/components/schemas/BankAccountDTO"},"type":"array","title":"Bank Accounts"}},"type":"object","required":["id","login","hashed_password","created_at","updated_at","deleted_at"],"title":"UserDTO"},"UserUpdateDTO":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"login":{"anyOf":[{"type":"string","maxLength":20,"minLength":4},{"type":"null"}],"title":"Login","description":"login must be without spaces and can contain only letters and numbers"},"password":{"anyOf":[{"type":"string","maxLength":20,"minLength":6},{"type":"null"}],"title":"Password","description":"password must be without spaces and can contain only letters and numbers"},"hashed_password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Hashed Password"},"deleted_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["deleted_at"],"title":"UserUpdateDTO"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}